Add configurable timeout and retry count for docker pushes.
By default we now try to push the image 3 times with a wait time of
5 seconds between tries.

This is a workaround for the docker registry returning HTTP 500 errors
when you push more than one image simultaneously to the same repository.
While this is mostly killing the symptoms of the actual problem it gives
us a better user experience, especially in continuous delivery pipelines
where the likelihood of pushing to the same repository is relatively high.